requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
colorama==0.4.4
coverage==6.3.3
cycler==0.11.0
django-appconf==1.0.5
django-compressor==4.0
django-filter==21.1
django-libsass==0.9
django-pwa==1.0.10
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djlint==0.7.6
fonttools==4.28.5
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
kiwisolver==1.3.2
libsass==0.21.0
matplotlib==3.5.1
numpy==1.21.5
opencv-python==4.5.5.64
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzbar==0.1.8
rcssmin==1.1.0
regex==2022.3.15
requests==2.27.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomlkit==0.9.2
tqdm==4.63.1
tzdata==2021.5
urllib3==1.26.9
zipp==3.7.0
